requirements.txt
python-dotenv
groq
PyPDF2
gradio